% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanmydata.R, R/cleanmydata_doc.R
\name{cleanmydata}
\alias{cleanmydata}
\title{Basic cleaning of data}
\usage{
cleanmydata(mydf, factorVars, numericVars, drop_na = TRUE)

cleanmydata(mydf, factorVars, numericVars, drop_na = TRUE)
}
\arguments{
\item{mydf}{use cbind.data.frame() to create sub-dataset with your variables of interest. Specify that 'stringAsFactors=FALSE"}

\item{factorVars}{variables you want to convert to factors}

\item{numericVars}{variables you want to convert to numeric}

\item{drop_na}{true or false variables if you want to drop missing variables or not}

\item{mydf}{use cbind.data.frame() to create sub-dataset with your variables of interest. Specify that 'stringAsFactors=FALSE"}

\item{factorVars}{variables you want to convert to factors}

\item{numericVars}{variables you want to convert to numeric}

\item{drop_na}{true or false variables if you want to drop missing variables or not}
}
\value{
new dataframe with cleaned variables

new dataframe with cleaned variables
}
\description{
Basic cleaning of data

Basic cleaning of data
}

